#3218 Synoptic panel file problem: - Moved NettyTerminateUtils to impl package; - Added PollingDataSourceUtils.terminateNetty;
