+++ src/com/serotonin/mango/MangoContextListener.java
@@ -68,7 +68,7 @@ import org.scada_lts.cache.PointHierarchyCache;
 import org.scada_lts.cache.ViewHierarchyCache;
 import org.scada_lts.config.ScadaVersion;
 import org.scada_lts.dao.SystemSettingsDAO;
-import org.scada_lts.ds.polling.protocol.opcua.client.NettyTerminateUtils;
+import org.scada_lts.ds.polling.PollingDataSourceUtils;
 import org.scada_lts.mango.adapter.MangoScadaConfig;
 import org.scada_lts.quartz.EverySecond;
 import org.scada_lts.quartz.EverySecondTool;
@@ -237,7 +237,7 @@ public class MangoContextListener implements ServletContextListener {
 		utilitiesTerminate(ctx);
 		highPriorityServiceTerminate();
 		databaseTerminate(ctx);
-		nettyTerminate();
+		terminatePollingDataSource();
 
 		Common.ctx = null;
 
@@ -741,7 +741,7 @@ public class MangoContextListener implements ServletContextListener {
 		return standardContext.getManager();
 	}
 
-	private static void nettyTerminate() {
-		NettyTerminateUtils.releaseSharedResources();
+	private static void terminatePollingDataSource() {
+		PollingDataSourceUtils.terminateNetty();
 	}
 }
